Refactor SearchPage
Update import paths and reorganize components
