+++ src/pages/SearchPage/components/SearchResult/index.ts
