+++ src/pages/SearchPage/store/userSearch.store.ts
